backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
SpeechRecognition==3.10.0
openai-whisper==20231106
torch==2.0.1
torchaudio==2.0.2
sounddevice==0.4.6
numpy==1.26.0
python-multipart==0.0.6
pyaudio==0.2.13
vosk